requirements.txt
numpy
pandas
kafka-python
xgboost
sqlalchemy
scikit-learn